+++ cpu-tracker.h
@@ -7,24 +7,25 @@
 #include <unistd.h>
 #include <signal.h>
 #include <string.h>
+#include <stdint.h>
 
 #define BUFFER_SIZE 10
 
 typedef struct CPU_DATA {
-    int cpu;
-    unsigned long user,
-                  nice,
-                  system,
-                  idle,
-                  iowait,
-                  irq,
-                  softirq,
-                  steal,
-                  guest,
-                  guest_nice;
+    int32_t cpu;
+    uint64_t    user,
+                nice,
+                system,
+                idle,
+                iowait,
+                irq,
+                softirq,
+                steal,
+                guest,
+                guest_nice;
 } CPU_DATA;
 
-extern int onlineProcessorsAmount;
+extern uint8_t onlineProcessorsAmount;
 
 extern pthread_mutex_t readBufferMutex;
 extern sem_t readSemaphoreEmpty;
@@ -35,4 +36,4 @@ extern sem_t printSemaphoreEmpty;
 extern sem_t printSemaphoreFull;
 
 CPU_DATA *get_readbuffer_data(void);
-unsigned long *get_printbuffer_data(void);
+uint64_t *get_printbuffer_data(void);
